Manage user-defined column width state in table component
