+++ src/client/src/components/table/table.tsx
@@ -1,4 +1,4 @@
-import { FC, MutableRefObject } from "react";
+import { FC, MutableRefObject, useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import {
   DataGrid,
@@ -11,6 +11,7 @@ import {
   GridRowSelectionModel,
   GridSortModel,
   GridToolbar,
+  useGridApiRef,
 } from "@mui/x-data-grid";
 import { GridApiCommunity } from "@mui/x-data-grid/internals";
 import { BoreholeAttributes } from "../../api-lib/ReduxStateInterfaces.ts";
@@ -42,7 +43,6 @@ interface TableProps {
   isDisabled?: boolean;
   showQuickFilter?: boolean;
   rowAutoHeight?: boolean;
-  onColumnResize?: (params: GridColumnResizeParams) => void;
 }
 
 export const Table: FC<TableProps> = ({
@@ -68,9 +68,31 @@ export const Table: FC<TableProps> = ({
   isDisabled = false,
   showQuickFilter = true,
   rowAutoHeight = false,
-  onColumnResize,
 }) => {
   const { i18n } = useTranslation();
+  const [columnWidths, setColumnWidths] = useState<Record<string, number>>({});
+  const internalApiRef = useGridApiRef();
+  const effectiveApiRef = apiRef ?? internalApiRef;
+
+  // Ensure user-defined column widths are persisted when table re-renders
+  useEffect(() => {
+    if (effectiveApiRef.current) {
+      Object.entries(columnWidths).forEach(([field, width]) => {
+        effectiveApiRef.current.setColumnWidth(field, width);
+      });
+    }
+  }, [apiRef, columnWidths, effectiveApiRef]);
+
+  const handleColumnResize = useCallback(
+    (params: GridColumnResizeParams) => {
+      const updatedWidths = {
+        ...columnWidths,
+        [params.colDef.field]: params.width,
+      };
+      setColumnWidths(updatedWidths);
+    },
+    [columnWidths],
+  );
 
   const disabledStyles = {
     cursor: isDisabled ? "default" : "pointer",
@@ -79,6 +101,20 @@ export const Table: FC<TableProps> = ({
   };
 
   const defaultRowClassName = (params: GridRowParams): string => (params.row.isDisabled ? "disabled-row" : "");
+
+  // Apply user-defined column widths to columns
+  const adjustedWidthColumns = columns.map(col => {
+    // If column is not resizable, no need to set width
+    if (col.resizable === false) {
+      return col;
+    }
+
+    return {
+      ...col,
+      width: columnWidths[col.field] || col.width,
+      flex: columnWidths[col.field] ? undefined : 1, // Auto flex if no width
+    };
+  });
 
   return (
     <DataGrid
@@ -89,7 +125,7 @@ export const Table: FC<TableProps> = ({
       loading={isLoading ?? !rows?.length}
       rowCount={rowCount ?? rows?.length}
       rows={rows}
-      columns={columns}
+      columns={adjustedWidthColumns}
       getRowHeight={() => (rowAutoHeight ? "auto" : 44)}
       onRowClick={onRowClick}
       pageSizeOptions={[100]}
@@ -119,7 +155,7 @@ export const Table: FC<TableProps> = ({
       onSortModelChange={onSortModelChange}
       paginationModel={paginationModel}
       onPaginationModelChange={onPaginationModelChange}
-      onColumnResize={onColumnResize}
+      onColumnResize={handleColumnResize}
       apiRef={apiRef}
       paginationMode={paginationMode}
       sortingMode={sortingMode}
